remove all logic from App component
